Se creo una clase principal para reutilizar metodos a traves de la herencia
